[BE] 일정, 일정날짜(EventDate) Create, Update, Delete API 구현 (#236)
* feat: 일정 날짜 추가, 삭제 api 구현

* feat: Event 추가, 수정, 삭제 api 구현

* refactor: 네이밍 컨벤션  수정

* refactor: 줄바꿈 컨벤션 수정

* feat: EventDate 삭제시 관련 Event 삭제하는 기능 추가

* refactor: 주석 삭제

* refactor: 줄바꿈 컨벤션 수정

* refactor: 메서드 위치 컨벤션 수정

* refactor: 주석 설명 요약

* refactor: EventDateRequestFixture 도입

* refactor: EventRequestFixture 도입

* refactor: 스웨거 주석 설명 구체화

일정 날짜 -> 행사일, 일정 -> 이벤트

* refactor: Request 스웨거 명세서 추가

* refactor: Event, EventDate 클래스 분리

* refactor: 미사용 메서드 제거

* refactor: TODO 주석 삭제

* refactor: 미사용 필드 삭제

* refactor: Request Fixture 사용

* refactor: RequestParam 사용하도록 수정

* refactor: 스웨거 주석 설명 구체화

* refactor: 메서드 네이밍 컨벤션

* refactor: 스웨거 주석 설명 구체화

* refactor: delete 메서드 네이밍 컨벤션 수정

* refactor: 스웨거 주석 설명 구체화

* refactor: 컨트롤러 미사용 응답 제거

* refactor: 패키지 이름 변경

* refactor: api 엔드포인트 수정

* refactor: api 엔드포인트 수정

* refactor: 예외 발생시 롤백을 위한 트랙잭션 추가

* refactor: 스웨거 주석 변경

* refactor: 스웨거 주석 변경

* refactor: 반환값에 따른 응답 코드 수정

* refactor: 메서드 네이밍 컨벤션에 맞게 수정

* refactor: setFixedClock Fixture 값으로 변경

* refactor: 다중 검증 assertSoftly 사용하도록 수정

* feat: 일정 날짜 수정 API 추가

* refactor: 스웨거 주석 변경

* refactor: 테스트 사전 데이터 jpa로 삽입하도록 수정

* refactor: 매직 넘버 변수로 추출

* refactor: 스웨거 주석 변경
