+++ backend/src/main/java/com/daedan/festabook/event/dto/EventDateResponse.java
@@ -1,0 +1,17 @@
+package com.daedan.festabook.event.dto;
+
+import com.daedan.festabook.event.domain.EventDate;
+import java.time.LocalDate;
+
+public record EventDateResponse(
+        Long id,
+        LocalDate date
+) {
+
+    public static EventDateResponse from(EventDate eventDate) {
+        return new EventDateResponse(
+                eventDate.getId(),
+                eventDate.getDate()
+        );
+    }
+}
