--- backend/src/main/java/com/daedan/festabook/schedule/domain/Event.java
@@ -1,66 +1,0 @@
-package com.daedan.festabook.schedule.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import java.time.Clock;
-import java.time.LocalTime;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Event implements Comparable<Event> {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private LocalTime startTime;
-
-    private LocalTime endTime;
-
-    @Column(nullable = false)
-    private String title;
-
-    @Column(nullable = false)
-    private String location;
-
-    @JoinColumn(nullable = false)
-    @ManyToOne(fetch = FetchType.LAZY)
-    private EventDate eventDate;
-
-    public Event(
-            LocalTime startTime,
-            LocalTime endTime,
-            String title,
-            String location,
-            EventDate eventDate
-    ) {
-        this.startTime = startTime;
-        this.endTime = endTime;
-        this.title = title;
-        this.location = location;
-        this.eventDate = eventDate;
-    }
-
-    public EventStatus determineStatus(Clock clock) {
-        return EventStatus.determine(clock, eventDate.getDate(), startTime, endTime);
-    }
-
-    @Override
-    public int compareTo(Event otherEvent) {
-        if (startTime.equals(otherEvent.startTime)) {
-            return endTime.compareTo(otherEvent.endTime);
-        }
-        return startTime.compareTo(otherEvent.startTime);
-    }
-}
