--- backend/src/main/java/com/daedan/festabook/schedule/dto/EventDateResponse.java
@@ -1,17 +1,0 @@
-package com.daedan.festabook.schedule.dto;
-
-import com.daedan.festabook.schedule.domain.EventDate;
-import java.time.LocalDate;
-
-public record EventDateResponse(
-        Long id,
-        LocalDate date
-) {
-
-    public static EventDateResponse from(EventDate eventDate) {
-        return new EventDateResponse(
-                eventDate.getId(),
-                eventDate.getDate()
-        );
-    }
-}
