--- backend/src/main/java/com/daedan/festabook/schedule/dto/EventResponses.java
@@ -1,19 +1,0 @@
-package com.daedan.festabook.schedule.dto;
-
-import com.daedan.festabook.schedule.domain.Event;
-import com.fasterxml.jackson.annotation.JsonValue;
-import java.time.Clock;
-import java.util.List;
-
-public record EventResponses(
-        @JsonValue List<EventResponse> responses
-) {
-
-    public static EventResponses from(List<Event> events, Clock clock) {
-        return new EventResponses(
-                events.stream()
-                        .map(event -> EventResponse.from(event, clock))
-                        .toList()
-        );
-    }
-}
