--- backend/src/test/java/com/daedan/festabook/schedule/domain/EventTest.java
@@ -1,52 +1,0 @@
-package com.daedan.festabook.schedule.domain;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.time.Clock;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.ZoneId;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class EventTest {
-
-    @Nested
-    class determineStatus {
-        
-        @ParameterizedTest(name = "날짜: {0}, 시작 시간: {1}, 종료 시간: {2}, 결과: {3}")
-        @CsvSource({
-                "2025-05-04, 10:00, 12:00, COMPLETED",   // 종료
-                "2025-05-05, 14:00, 17:00, ONGOING",     // 진행중
-                "2025-05-05, 17:00, 18:00, UPCOMING",    // 예정
-        })
-        void 이벤트_상태_판별_파라미터_테스트(LocalDate date, LocalTime startTime, LocalTime endTime, EventStatus expected) {
-            // given
-            ZoneId korea = ZoneId.of("Asia/Seoul");
-            Clock clock = Clock.fixed(
-                    LocalDateTime.of(2025, 5, 5, 16, 0).atZone(korea).toInstant(),
-                    korea
-            );
-
-            Event event = EventFixture.create(
-                    startTime,
-                    endTime,
-                    EventDateFixture.create(date)
-            );
-
-            // when
-            EventStatus result = event.determineStatus(clock);
-
-            // then
-            assertThat(result).isEqualTo(expected);
-        }
-    }
-}
